[REFACTOR] create Form into its own component
