+++ time-capsule/src/components/CreateForm/CreateForm.module.css
@@ -1,0 +1,15 @@
+.content {
+  padding: 1rem;
+  display: grid;
+  gap: 1.5rem;
+
+  .title {
+    font-size: 1.5rem;
+    font-weight: 500;
+  }
+
+  .action {
+    display: flex;
+    justify-content: space-between;
+  }
+}
