+++ time-capsule/src/components/CreateModal/CreateModal.module.css
@@ -15,20 +15,4 @@
   &::backdrop {
     backdrop-filter: blur(2px);
   }
-
-  .content {
-    padding: 1rem;
-    display: grid;
-    gap: 1.5rem;
-
-    .title {
-      font-size: 1.5rem;
-      font-weight: 500;
-    }
-
-    .action {
-      display: flex;
-      justify-content: space-between;
-    }
-  }
 }
