+++ time-capsule/src/components/CreateModal/CreateModal.tsx
@@ -1,17 +1,14 @@
-import { FormEvent, ReactNode, useRef } from "react";
+import { ReactNode, useRef } from "react";
 
 import { Capsule } from "../../types/capsule.ts";
 
 import Button from "../Button/button.tsx";
-import TextInput from "../TextInput/TextInput.tsx";
-import TeaxtArea from "../TextArea/TeaxtArea.tsx";
-import DatetimeLocl from "../DatetimeLocal/DatetimeLocal.tsx";
-import Select from "../Select/Select.tsx";
 
 import MdiLightPlus from "../../icons/MdiLightPlus.tsx";
 
 import styles from "./CreateModal.module.css";
-import { Category } from "../../types/category.ts";
+
+import CreateForm from "../CreateForm/CreateForm.tsx";
 
 type Props = {
   setCpsule: React.Dispatch<React.SetStateAction<Capsule[]>>;
@@ -23,26 +20,6 @@ function CreateModal({ setCpsule }: Props): ReactNode {
   const dialogHandler = (): void => {
     dialogRef.current?.showModal();
   };
-  const CancelHandler = (): void => {
-    dialogRef.current?.close();
-  };
-
-  const submitHandler = (e: FormEvent<HTMLFormElement>): void => {
-    e.preventDefault();
-
-    const formData = new FormData(e.currentTarget);
-
-    const capsule: Capsule = {
-      id: 5,
-      title: formData.get("title") as string,
-      describtion: formData.get("describtion") as string,
-      date: new Date(formData.get("date") as string),
-      category: formData.get("category") as Category,
-    };
-
-    setCpsule((old) => [...old, capsule]);
-    dialogRef.current?.close();
-  };
 
   return (
     <>
@@ -53,29 +30,11 @@ function CreateModal({ setCpsule }: Props): ReactNode {
       </div>
 
       <dialog className={styles.dialog} ref={dialogRef}>
-        <form className={styles.content} onSubmit={submitHandler}>
-          <div className={styles.title}>New Appointment</div>
-
-          <TextInput name="title" placeholder="Youre title..."></TextInput>
-          <TeaxtArea name="description" placeholder="Youre description" />
-          <DatetimeLocl name="date" />
-          <Select
-            name="category"
-            variant={"outlined"}
-            options={[
-              { value: "work", label: "work" },
-              { value: "friendly", label: "friendly" },
-              { value: "family", label: "family" },
-            ]}
-          ></Select>
-
-          <div className={styles.action}>
-            <Button type={"button"} variant={"solid"} onClick={CancelHandler}>
-              Cancel
-            </Button>
-            <Button variant={"otlined"}>Apply</Button>
-          </div>
-        </form>
+        <CreateForm
+          onCancle={() => dialogRef.current?.close()}
+          onSubmit={() => dialogRef.current?.close()}
+          setCpsule={setCpsule}
+        />
       </dialog>
     </>
   );
